# api/src/generated/openapi.yaml
openapi: 3.0.0
components:
...
  schemas:
    User:
      description: User model - represents a user in the system
      properties:
        id:
          type: string
          description: Unique user identifier
        name:
          type: string
          description: User's name
        displayName:
          type: string
          description: User's display name with location and year
        avatar:
          type: string
          nullable: true
          description: DiceBear bottts avatar seed
        createdAt:
          type: string
          format: date-time
          description: Timestamp when the user was created
        updatedAt:
          type: string
          format: date-time
          description: Timestamp when the user was last updated
      required:
        - id
        - name
        - displayName
        - createdAt
        - updatedAt
      type: object
      additionalProperties: false
    UserInput:
      description: Input for creating a new user
      properties:
        id:
          type: string
        name:
          type: string
        displayName:
          type: string
        avatar:
          type: string
          nullable: true
      required:
        - id
        - name
        - displayName
      type: object
      additionalProperties: false
    UserUpdate:
      description: Input for updating a user
      properties:
        name:
          type: string
        displayName:
          type: string
        avatar:
          type: string
          nullable: true
      type: object
      additionalProperties: false
...
  securitySchemes: {}
info:
  title: BoltUp API
  version: 1.0.0
  description: REST API for BoltUp - A social event planning platform for groups
  license:
    name: ISC
  contact:
    name: BoltUp Team
paths:
  /users:
    get:
      operationId: GetUsers
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/User"
                type: array
      description: Get all users
      summary: Retrieves a list of all users in the system
      tags:
        - Users
      security: []
      parameters: []
    post:
      operationId: CreateUser
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      description: Create a new user
      summary: Creates a new user in the system
      tags:
        - Users
      security: []
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
  /users/{id}:
    get:
      operationId: GetUser
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
      description: Get user by ID
      summary: Retrieves a single user by their unique identifier
      tags:
        - Users
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
    put:
      operationId: UpdateUser
      responses:
        "200":
          description: Ok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      description: Update a user
      summary: Updates an existing user's information
      tags:
        - Users
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
    delete:
      operationId: DeleteUser
      responses:
        "204":
          description: No Content
      description: Delete a user
      summary: Deletes a user from the system
      tags:
        - Users
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
...
servers:
  - url: http://localhost:3000/api